requirements.txt
pyrogram
python-telegram-bot
python-dotenv
logging
asyncio